+++ xilinx/src/xlnx_scaler.rs
@@ -109,3 +109,86 @@ impl Drop for XlnxScaler {
         }
     }
 }
+
+#[cfg(test)]
+mod scaler_tests {
+    use crate::{tests::*, xlnx_scal_props::*, xlnx_scal_utils::*, xlnx_scaler::*};
+
+    #[test]
+    fn test_abr_scale() {
+        let mut frame_props = XmaFrameProperties {
+            format: XmaFormatType_XMA_YUV420_FMT_TYPE,
+            width: 1280,
+            height: 720,
+            bits_per_pixel: 8,
+            ..Default::default()
+        };
+        // read the raw yuv into xma_frames, This file has 300 frames.
+        let (xma_frames, _y_component, _x_component) = read_raw_yuv_420_p(RAW_YUV_P_FILE_PATH, &mut frame_props, 300);
+
+        initialize();
+
+        // create a Xlnx scalar's context
+        let mut scal_props = Box::new(XlnxScalerProperties {
+            in_width: 1280,
+            in_height: 720,
+            fr_num: 1,
+            fr_den: 25,
+            nb_outputs: 3,
+            out_width: [1280, 852, 640, 0, 0, 0, 0, 0],
+            out_height: [720, 480, 360, 0, 0, 0, 0, 0],
+            enable_pipeline: 0,
+            log_level: 3,
+            latency_logging: 1,
+        });
+
+        let scal_params: [XmaParameter; MAX_SCAL_PARAMS] = Default::default();
+        let mut scal_params = Box::new(scal_params);
+
+        let mut xma_scal_props = xlnx_create_xma_scal_props(&mut scal_props, &mut scal_params).unwrap();
+
+        let xrm_ctx = unsafe { xrmCreateContext(XRM_API_VERSION_1) };
+
+        let cu_res: xrmCuResource = Default::default();
+
+        let mut xlnx_scal_ctx = XlnxScalerXrmCtx {
+            xrm_reserve_id: 0,
+            device_id: -1,
+            scal_load: xlnx_calc_scal_load(xrm_ctx, &mut *xma_scal_props).unwrap(),
+            scal_res_in_use: false,
+            xrm_ctx,
+            num_outputs: 3,
+            cu_res,
+        };
+
+        // create xlnx scaler
+        let mut scaler = XlnxScaler::new(&mut *xma_scal_props, &mut xlnx_scal_ctx).unwrap();
+
+        let mut processed_frame_count = 0;
+
+        for xma_frame in xma_frames {
+            match scaler.process_frame(xma_frame) {
+                Ok(_) => {
+                    // successfully scaled frame.
+                    processed_frame_count += 1;
+                    // clear xvbm buffers for each return frame
+                    for i in 0..xma_scal_props.num_outputs as usize {
+                        unsafe {
+                            let handle: XvbmBufferHandle = (*scaler.out_frame_list[i]).data[0].buffer;
+                            xvbm_buffer_pool_entry_free(handle);
+                        }
+                    }
+                }
+                Err(e) => match e.err {
+                    XlnxErrorType::XlnxSendMoreData => {}
+                    _ => panic!("scalar processing has failed with error {:?}", e),
+                },
+            };
+            if !xma_frame.is_null() {
+                unsafe { xma_frame_free(xma_frame) };
+            }
+        }
+
+        assert_eq!(processed_frame_count, 300);
+    }
+}
